Bug fixes, some optimizations
1. 32/64-bit ASIO driver and channel selection is stored separately.
2. Optimized listbox drawing.
3. Fixed 64-bit related warnings.

+++ driver/winmm_drv.cpp
@@ -310,7 +310,7 @@ STDAPI_(DWORD) modMessage(DWORD uDeviceID, DWORD uMsg, DWORD_PTR dwUser, DWORD_P
 		return MMSYSERR_NOERROR;	
 		
 	case MODM_SETVOLUME:
-		driver->volume = dwParam1;	 
+		driver->volume = (DWORD)dwParam1;	 
 		midiSynth.SetVolume(uDeviceID, sqrt(float(LOWORD(dwParam1)) / 65535.f)); //falco: for separate port A/B note velocity
 		return MMSYSERR_NOERROR;
